move out hw_uart.h
