+++ main/constants.h
@@ -1,0 +1,11 @@
+/**
+ * constants.h
+ * 
+ */
+
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+
+#define MSG_LEN 512
+
+#endif
